requirements.txt
opencv-python
numpy
scipy
pillow
pytesseract
pdf2image

# sudo apt-get install tessaract-ocr
